% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampelatorFunctions.R
\name{adjustSampleSize}
\alias{adjustSampleSize}
\title{Adjust sample size value for finite population or expected missingness}
\usage{
adjustSampleSize(sampleSize, populationSize = NA,
  adjustFinitePopulation = FALSE, inflationFactor = NA)
}
\arguments{
\item{sampleSize}{numeric, the uncorrected estimated sample size}

\item{populationSize}{numeric, the population size; default is NA}

\item{adjustFinitePopulation}{boolean, adjust for finite population?; 
default is FALSE}

\item{inflationFactor}{numeric, the inflation factor with which the uncorrected
sample size should be multiplied to account e.g. for missingness;
default is NA}
}
\value{
numeric, the corrected estimated sample size
}
\description{
Adjust sample size value for finite population or expected missingness
}

