% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribessFunctions.R
\name{calculatePfree}
\alias{calculatePfree}
\title{Calculate the probability of freedom from disease}
\usage{
calculatePfree(targetPFree, pFreePrior, sse, pIntro, timePoints = 2)
}
\arguments{
\item{targetPFree}{a numeric, the targeted probability 
of freedom from disease}

\item{pFreePrior}{a numeric, the prior probability of freedom from disease}

\item{sse}{a numeric, the initial system sensitivity}

\item{pIntro}{a numeric vector, the probability that the disease was 
introduced since last survey or a given timespan; 
its length should correspond to the integer provided for timePoints}

\item{timePoints}{an integer, the number of time points for which the 
probability of freedom should be calculated; min value is 2}
}
\value{
data frame with named numeric vectors for: 
pFreePrior, the prior probability of freedom from disease;
pFree, the probability of freedom from disease;
sse, the system sensitivity;
pIntro, the probability that disease was introduced;
pFreeAdjusted, the probability of freedom from disease adjusted for the 
fact that disease is introduced with probability pIntro
}
\description{
Calculate the probability of freedom from disease
}

