% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectingSamples.R
\name{collapseData}
\alias{collapseData}
\title{Collapse data according to the user-defined variable}
\usage{
collapseData(inputData, collapseVariable)
}
\arguments{
\item{inputData}{data frame, the original data}

\item{collapseVariable}{character, variable name in inputData with respect
to which the data should be collapsed}
}
\value{
data frame, cbind (1) the unique rows of the inputData excluding 
collapseVariable, (2) the number of registrations per level of collapseVariable
 (instead of the column for collapseVariable) and (3) the total number of 
registrations
}
\description{
Collapse data according to the user-defined variable
}

