% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribessFunctions.R
\name{createRandom}
\alias{createRandom}
\title{Generate random values for uncertain input parameter}
\usage{
createRandom(distributionName, distributionParameters, isQuantiles = FALSE,
  nMC = 100, seed = 1)
}
\arguments{
\item{distributionName}{character, distribution that is assumed;
 one of the standard distribution names 
"beta", "cauchy", "chisq", "exp", "f", "gamma", "hyper", "lnorm", "nbinom",
"norm", "pois", "t", "unif", "weibull"; or "constant" if no uncertainty}

\item{distributionParameters}{numeric vector with named arguments 
needed for each of the distribution functions; 
for distribution "constant" provide the 'value'}

\item{isQuantiles}{boolean vector with same length as 
distributionNames; indicates whether quantiles and corresponding 
probabilities are provided in distributionParameters}

\item{nMC}{number of Monte Carlo experiments; default is 100}

\item{seed}{integer, seed for random number generation; default is 1}
}
\value{
matrix with nMC random values (1 column)
}
\description{
Generate random values for uncertain input parameter
}

