% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribessFunctions.R
\name{expandRisks}
\alias{expandRisks}
\title{Expand the all input list elements to all possible risk group combinations}
\usage{
expandRisks(unexpanded)
}
\arguments{
\item{unexpanded}{named list of risk factors that contain numeric vectors 
of e.g. relativeRisks per risk factor}
}
\value{
a dataframe with combined values for all possible risk group combinations
}
\description{
Expand the all input list elements to all possible risk group combinations
}

