% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribessFunctions.R
\name{getBestDistribution}
\alias{getBestDistribution}
\title{Get the best distributional fit & its parameters for known quantiles}
\usage{
getBestDistribution(probabilities, quantiles, withGUI = FALSE)
}
\arguments{
\item{probabilities}{a numeric vector of probabilities}

\item{quantiles}{a numeric vector of quantiles corresponding to probabilities}

\item{withGUI}{boolean, if TRUE a user-interface pops-up to choose the best
distributional shape, if FALSE automatically the best distribution is chosen
out of "normal", "beta" and "gamma"; default is FALSE}
}
\value{
a list with "bestName" the name of the best distribution function,
"bestParameters" a vector with the distribution parameter values and
"logLikelihood" a numeric, the (largest) log-likelihood value
}
\description{
Get the best distributional fit & its parameters for known quantiles
}

