% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribessFunctions.R
\name{getConvenience}
\alias{getConvenience}
\title{Calculate the sample size and group sensitivity
 for a given target value of the area system sensitivity
 while accounting for the convenience matrix}
\usage{
getConvenience(targetAse, inputConvenience, maxn = NA, isSampledMaxn = TRUE,
  totalN = NULL, dp, tse, method = c("hyper", "binom"),
  isRiskbased = FALSE, relativeRisks = NULL, proportionsRiskGroups = NULL)
}
\arguments{
\item{targetAse}{target value of the area system sensitivity}

\item{inputConvenience}{numeric vector with convenience values per risk factor}

\item{maxn}{maximum number of observations that can be sampled 
in the least convenient risk groups; default value is NA}

\item{isSampledMaxn}{boolean value, default is TRUE. Only used if a 
non-missing value for maxn is provided; 
if TRUE, sample size in risk groups with convenience 1 is automatically maxn,
if FALSE their sample size may be smaller but at most maxn}

\item{totalN}{total population size, default value is NULL}

\item{dp}{design prevalence}

\item{tse}{test sensitivity}

\item{method}{method of sample size calculation; 
one of 'hyper' (default) or 'binom'}

\item{isRiskbased}{logical Is the sampling risk based; 
default value is FALSE}

\item{relativeRisks}{a dataframe with relative risks for all possible 
risk group combinations; default value is NULL}

\item{proportionsRiskGroups}{a dataframe with relative proportions of 
total population for all possible risk group combinations; 
default value is NULL}
}
\value{
data frame reports for each risk group the population size, 
sample size and group sensitivity; such that the area sensitivity is targetAse
and size are according to input convenience values
}
\description{
Calculate the sample size and group sensitivity
 for a given target value of the area system sensitivity
 while accounting for the convenience matrix
}

