% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampelatorFunctions.R
\name{getDesignEffect}
\alias{getDesignEffect}
\title{Calculate the design effect}
\usage{
getDesignEffect(clusterVariance, conditionalPrevalence = NA,
  correlation = NA, clusterSize)
}
\arguments{
\item{clusterVariance}{numeric, estimated variance between the cluster means}

\item{conditionalPrevalence}{numeric, conditional prevalence of the outcome
given a random effect of zero; default is NA}

\item{correlation}{numeric, the intra-class correlation; default is NA
if NA correlation is estimated based on approximation using clusterVariance
and clusterSize}

\item{clusterSize}{integer, (average) size of the clusters}
}
\value{
numeric, the design effect
}
\description{
Calculate the design effect
}

