% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampelatorFunctions.R
\name{getDesiredDifference}
\alias{getDesiredDifference}
\title{Calculate the desired difference for three stage sampling design}
\usage{
getDesiredDifference(varianceLevel3, varianceLevel2, marginalPrevalence0 = NA,
  marginalPrevalenceA = NA, conditionalPrevalence0,
  conditionalPrevalenceA = NA)
}
\arguments{
\item{varianceLevel3}{numeric, the variance between level 3 units (highest level)}

\item{varianceLevel2}{numeric, the variance between level 2 units (middle level)}

\item{marginalPrevalence0}{numeric, marginal prevalence under the 
null hypothesis; default is NA}

\item{marginalPrevalenceA}{numeric, marginal prevalence under the 
alternative hypothesis; default is NA}

\item{conditionalPrevalence0}{numeric, conditional prevalence under the 
null hypothesis; default is NA}

\item{conditionalPrevalenceA}{numeric, conditional prevalence under the 
alternative hypothesis; default is NA}
}
\value{
dataframe with desired difference, randomEffect0 i.e. conditional 
prevalence on the logodds scale under the null hypothesis, randomEffectA i.e.
conditional prevalence on the logodds scale under the alternative hypothesis
}
\description{
Calculate the desired difference for three stage sampling design
}

