% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampelatorFunctions.R
\name{getMissing}
\alias{getMissing}
\title{Check which of the input parameters is missing}
\usage{
getMissing(purpose = c("estimation", "testing"), inputList)
}
\arguments{
\item{purpose}{character string, the purpose of the study is one of 
"estimation" or "testing"}

\item{inputList}{list with input parameters}
}
\value{
boolean vector, which of the variables in inputList is missing;
error message if not exactly one necessary parameter of inputList is missing
}
\description{
Check which of the input parameters is missing
}

