% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampelatorFunctions.R
\name{getRawCalculation}
\alias{getRawCalculation}
\title{Calculate (unrounded) sample size, desired difference or statistical power
using a general formula}
\usage{
getRawCalculation(purpose = c("estimation", "testing"), sampleSize = NA,
  desiredDifference = NA, power = NA, typeIerror, variance,
  populationSize = NA, adjustFinitePopulation = FALSE,
  inflationFactor = NA)
}
\arguments{
\item{purpose}{character string, the purpose of the study is one of 
"estimation" or "testing"}

\item{sampleSize}{integer, the sample size; default is NA}

\item{desiredDifference}{numeric, for "estimation" the desired margin of error,
i.e. half width of the desired confidence interval or for "testing" the 
true difference in means that is tested; default is NA}

\item{power}{numeric, statistical power to detect the predefined difference 
"desiredDifference" when purpose is "testing"; default is NA}

\item{typeIerror}{numeric, the type I error}

\item{variance}{numeric, e.g. estimated element variance in simple random
sampling design}

\item{populationSize}{numeric, the population size; default is NA}

\item{adjustFinitePopulation}{boolean, adjust for finite population?; 
default is FALSE}

\item{inflationFactor}{numeric, the inflation factor with which the uncorrected
sample size should be multiplied to account e.g. for missingness;
default is NA}
}
\value{
a numeric; when purpose is estimation, one of sampleSize or 
desiredDifference is missing and the returned value; when purpose is testing,
one of sampleSize, desiredDifference or power is missing and the returned value
}
\description{
Calculate (unrounded) sample size, desired difference or statistical power
using a general formula
}

