% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampelatorFunctions.R
\name{getThreeStage}
\alias{getThreeStage}
\title{Calculate design statistics for three stage sampling}
\usage{
getThreeStage(purpose = c("estimation", "testing"), sampleSize = NA,
  desiredDifference = NA, power = NA, typeIerror,
  marginalPrevalence0 = NA, marginalPrevalenceA = NA,
  conditionalPrevalence0, conditionalPrevalenceA = NA, sizeLevel3 = NA,
  sizeLevel2 = NA, sizeLevel1 = NA, varianceLevel3, varianceLevel2,
  varianceLevel1 = NA, populationSize = NA,
  adjustFinitePopulation = FALSE, inflationFactor = NA)
}
\arguments{
\item{purpose}{character string, the purpose of the study is one of 
"estimation" or "testing"}

\item{sampleSize}{integer, the total sample size; default is NA}

\item{desiredDifference}{numeric, for "estimation" the desired margin of error,
i.e. half width of the desired confidence interval or for "testing" the 
true difference in means that is tested; default is NA}

\item{power}{numeric, statistical power to detect the predefined difference 
"desiredDifference" when purpose is "testing"; default is NA}

\item{typeIerror}{numeric, the type I error}

\item{marginalPrevalence0}{numeric, marginal prevalence under the 
null hypothesis; default is NA}

\item{marginalPrevalenceA}{numeric, marginal prevalence under the
alternative hypothesis; default is NA}

\item{conditionalPrevalence0}{numeric, conditional prevalence under the 
null hypothesis; default is NA}

\item{conditionalPrevalenceA}{numeric, conditional prevalence under the 
alternative hypothesis; default is NA}

\item{sizeLevel3}{integer, the number of level 3 units (highest level); default is NA}

\item{sizeLevel2}{integer, the number of level 2 units (middle level); default is NA}

\item{sizeLevel1}{integer, the number of level 1 units (lowest level); default is NA}

\item{varianceLevel3}{numeric, the variance between level 3 units (highest level)}

\item{varianceLevel2}{numeric, the variance between level 2 units (middle level)}

\item{varianceLevel1}{numeric, the variance between level 1 units (lowest level);
default is NA, if missing a value for conditionalPrevalenceA should be provided}

\item{populationSize}{numeric, the population size; default is NA}

\item{adjustFinitePopulation}{boolean, adjust for finite population?; 
default is FALSE}

\item{inflationFactor}{numeric, the inflation factor with which the uncorrected
sample size should be multiplied to account e.g. for missingness;
default is NA}
}
\value{
a dataframe with sampleSize, desiredDifference, power,
sizeLevel3, sizeLevel2 and sizeLevel1
}
\description{
Calculate design statistics for three stage sampling
}

