% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampelatorFunctions.R
\name{getTwoStage}
\alias{getTwoStage}
\title{Calculate design statistics and sample alllocation for 
clustered sampling design within strata}
\usage{
getTwoStage(purpose = c("estimation", "testing"), sampleSize = NA,
  desiredDifference = NA, power = NA, typeIerror,
  allocation = "proportional", clusterSize = NA, clusterVariances,
  stratumProportions, designEffect = NA, stratumVariances,
  populationSize = NA, adjustFinitePopulation = FALSE,
  inflationFactor = NA, exactProduct = TRUE, exactSum = TRUE)
}
\arguments{
\item{purpose}{character string, the purpose of the study is one of 
"estimation" or "testing"}

\item{sampleSize}{integer, the total sample size; default is NA}

\item{desiredDifference}{numeric, for "estimation" the desired margin of error,
i.e. half width of the desired confidence interval or for "testing" the 
true difference in means that is tested; default is NA}

\item{power}{numeric, statistical power to detect the predefined difference 
"desiredDifference" when purpose is "testing"; default is NA}

\item{typeIerror}{numeric, the type I error}

\item{allocation}{method to allocate the total sample size over the strata;
one of "proportional" or "neyman"}

\item{clusterSize}{integer, (average) size of the clusters}

\item{clusterVariances}{numeric vector, estimated variance between cluster means
within each stratum}

\item{stratumProportions}{numeric vector, the expected proportion of the 
population in each stratum}

\item{designEffect}{numeric vector, the design effect for each stratum,
i.e. the ratio of the stratum mean variances under multistage sampling and 
stratified sampling; default value is NA}

\item{stratumVariances}{numeric vector, estimated variance within each stratum}

\item{populationSize}{numeric, the population size; default is NA}

\item{adjustFinitePopulation}{boolean, adjust for finite population?; 
default is FALSE}

\item{inflationFactor}{numeric, the inflation factor with which the uncorrected
sample size should be multiplied to account e.g. for missingness;
default is NA}

\item{exactProduct}{boolean, if TRUE the sampleSize is exactly the product of
numberClusters and clusterSize; if FALSE the sampleSize may be smaller than
the product of numberClusters and clusterSize; default is TRUE}

\item{exactSum}{boolean, if TRUE the total sampleSize is exactly the sum of
all stratumSizes; if FALSE the sampleSize may be smaller than the the sum of 
all stratumSizes; default is TRUE}
}
\value{
a list with two dataframes: one with sampleAllocation info 
(stratumVariances, stratumProportions, availableClusters, numberClusters,
stratumSize); one with design statistics (numberClusters, sampleSize, 
desiredDifference, power)
}
\description{
Calculate design statistics and sample alllocation for 
clustered sampling design within strata
}

