% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampelatorFunctions.R
\name{getVarianceCovariance}
\alias{getVarianceCovariance}
\title{Construct variance-covariance matrix}
\usage{
getVarianceCovariance(varianceCorrelation, varianceSingle = NA,
  correlationSingle = NA, dimension = NA)
}
\arguments{
\item{varianceCorrelation}{numeric matrix with variance at the diagonal and
pairwise correlation values at non-diagonal}

\item{varianceSingle}{a single value for the variance; default is NA}

\item{correlationSingle}{a single value for the correlation; default is NA}

\item{dimension}{integer, determines the dimension (nrows = ncols) of the
output matrix if varianceSingle and correlationSingle are provided}
}
\value{
variance-covariance matrix
}
\description{
Construct variance-covariance matrix
}

