% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribessFunctions.R
\name{helpExpandNames}
\alias{helpExpandNames}
\title{Expand the input list to all possible risk group combinations}
\usage{
helpExpandNames(inputList, toPaste = TRUE)
}
\arguments{
\item{inputList}{inputList list of risk factors that contain numeric vectors 
of e.g. relativeRisks per risk factor}

\item{toPaste}{boolean; TRUE (default) then return 1 pasted string per 
combination, FALSE then return a data frame with columns the risk factors}
}
\value{
vector or data frame that combines names for all possible 
risk group combinations; a unique name per combination is given
}
\description{
Expand the input list to all possible risk group combinations
}

