% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribessFunctions.R
\name{helpExpandRisks}
\alias{helpExpandRisks}
\title{Expand the input list to all possible risk group combinations}
\usage{
helpExpandRisks(inputList)
}
\arguments{
\item{inputList}{list of risk factors that contain numeric vectors 
of e.g. relativeRisks per risk factor}
}
\value{
vector that combines values of inputList for all possible
risk group combinations
}
\description{
Expand the input list to all possible risk group combinations
}

