% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribessFunctions.R
\name{histogramWrapper}
\alias{histogramWrapper}
\title{Plot Histograms of Uncertainty}
\usage{
histogramWrapper(outputWrapper, varNames, timePeriods = NA)
}
\arguments{
\item{outputWrapper}{output from wrapFunction}

\item{varNames}{names of the variables for which to plot histograms}

\item{timePeriods, }{character vector with time periods for which the 
histogram should be plotted; default is NA}
}
\description{
Plot Histograms of Uncertainty
}

