% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampelatorFunctions.R
\name{roundUp}
\alias{roundUp}
\title{Take the ceiling of a numeric unless the tolerance level is not exceeded}
\usage{
roundUp(toRound, delta = 0.1)
}
\arguments{
\item{toRound}{numeric value that should be rounded}

\item{delta}{numeric; tolerance level; default is 0.1}
}
\value{
numeric value
}
\description{
Take the ceiling of a numeric unless the tolerance level is not exceeded
}

