% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectingSamples.R
\name{selectSample}
\alias{selectSample}
\title{Select a random sample from the data}
\usage{
selectSample(data, seed, size, weights = NA)
}
\arguments{
\item{data}{data frame from which the sample should be taken}

\item{seed}{integer, to assure reproducibility of the sampling}

\item{size}{integer, the size of the random sample}

\item{weights}{numeric vector, used for weighted sampling; default is NA}
}
\value{
data frame, random subsample of data
}
\description{
Select a random sample from the data
}

