% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectingSamples.R
\name{simulateSampling}
\alias{simulateSampling}
\title{Simulate the random sampling of data}
\usage{
simulateSampling(data, seed, size, nSimulations)
}
\arguments{
\item{data}{data frame from which the sample should be taken}

\item{seed}{integer, to assure reproducibility of the sampling}

\item{size}{integer, the size of the random sample}

\item{nSimulations}{the number of random samples that has to be drawn}
}
\value{
a list of data frames, each a random subsample of data
}
\description{
Simulate the random sampling of data
}

