% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectingSamples.R
\name{summarizeSampling}
\alias{summarizeSampling}
\title{Grid of histogram plots comparing simulated and selected samples}
\usage{
summarizeSampling(selectedSample, simulatedSamples, varName)
}
\arguments{
\item{selectedSample}{data frame, as given by selectSample()}

\item{simulatedSamples}{list of data frames, as given by simulateSampling()}

\item{varName}{character, the variable name for which histograms have to be 
plotted}
}
\value{
ggplot figure
}
\description{
Grid of histogram plots comparing simulated and selected samples
}

