% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribessFunctions.R
\name{summaryWrapper}
\alias{summaryWrapper}
\title{Select one scenario out of all Monte Carlo experiments}
\usage{
summaryWrapper(outputWrapper, percentile, parameterName)
}
\arguments{
\item{outputWrapper}{output object from the wrapperUncertain() function}

\item{percentile}{value between 0 and 1; the percentage of MC experiments with
smaller values for the parameterName than the one that will be reported}

\item{parameterName}{character; the selection will be based on this variable; 
if it consider several values per experiment (e.g. sampleSize per risk group) 
the sum will be taken}
}
\value{
Named list with results for the provided percentile
}
\description{
Select one scenario out of all Monte Carlo experiments
}

