% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribessFunctions.R
\name{wrapFunction}
\alias{wrapFunction}
\title{Wrap Monte Carlo experiments around one of the functions in this package 
to account for uncertainty on input parameters}
\usage{
wrapFunction(wrappedFunction = c("getSampleSizeGse", "optimizeSampleSizeGse",
  "getConvenience", "calculatePfree"), ..., nMC = 100)
}
\arguments{
\item{wrappedFunction}{function that needs to be wrapped 
One of "getSampleSizeGse", "optimizeSampleSizeGse", 
"getConvenience", "calculatePfree"}

\item{...}{further input parameters as passed to wrappedFunction}

\item{nMC}{number of Monte Carlo experiments}
}
\value{
Named list with output as from the original wrappedFunction for all 
Monte Carlo experiments
}
\description{
Wrap Monte Carlo experiments around one of the functions in this package 
to account for uncertainty on input parameters
}

