% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampelatorFunctions.R
\name{wrapFunctionMilanzi}
\alias{wrapFunctionMilanzi}
\title{Wrapper function around one of the functions in this package 
to account for uncertainty on some input parameters}
\usage{
wrapFunctionMilanzi(wrappedFunction = c("getSimpleRandom", "getDesignEffect",
  "getClustered", "getStratified", "getTwoStage", "getThreeStage",
  "getChangeOverTime"), randomInput = NULL, fixedInput = NULL)
}
\arguments{
\item{wrappedFunction}{function that needs to be wrapped}

\item{randomInput}{list of input parameters that are random; default is NULL}

\item{fixedInput}{list of input parameters that are non-random;
default is NULL}
}
\value{
a list with output elements as in the original wrappedFunction
}
\description{
Wrapper function around one of the functions in this package 
to account for uncertainty on some input parameters
}

